backend/requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
alembic==1.16.2
python-dotenv==1.1.1
feedparser==6.0.11
newsapi-python==0.2.7
openai==1.93.0
google-generativeai==0.8.5
schedule==1.2.2
requests==2.32.4 